Add a government-paas page
This is just about the simplest thing I could have done here, short of adding an
html page and doing some hand-coding.

All this does is add the page and hook into set-up around CSS so there is a new
page at /government-paas/ with no meaningful content, but styled in the same way
as the rest of the site.

It's very tempting to do some more clean-up around the whole site right now but
the aim here is just to do something simple for now.