Add new blog post and link to alphagov
A new blog post has been published since this PR was raised, so adding that.

Also, there previously was a commit that listed all our repos, but this was
manual and would require maintenance; and was already out of date. It makes more
sense to remove that and only add the repos when we have an automated way to
keep track of which is which.